feat: add an internal Command type for storing the plugin commands
