+++ main.go
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package main is the entry point for Reginald, the personal workstation valet.
-// TODO: Add a comment describing the actual command when there is something to
-// describe.
+/*
+Reginald is the personal workstation valet.
+
+TODO: Add a comment describing the actual command when there is something to
+describe.
+*/
 package main
 
 import (
@@ -33,6 +36,7 @@ func main() {
 	}
 }
 
+// run runs the CLI and returns the exit code.
 func run() int {
 	defer panichandler.Handle()
